Fix a bug which caused to show todo lists with outdated children
When updating the children of a todo list, the uuid of the todo list with be stored in DataService. When the user navigates to another page, the todo list will be updated. This is necessary because when updating the todo list immediatly after it was changed, the current todo list tree will be closed.