Fix some bugs
Fix a bug with updating the todos on the calendar page when adding todos to a todo list. Also fix a bug in the layout of the start page with showing todos and todo lists one after another. Fix another bug with old todo groups in the todo list modal.